refactor: single or multi download selection
